Removing topics in bundle check from ForwardingTopology and ASBTransportInfrastructure
Removing NamespaceBundleConfigurations from settings and making check static to return  ns bundle configuration directly.
Only ForwardingTopologySectionManager knows about NamespaceBundleConfigurations at this point.
