+++ src/Transport/Config/AzureServiceBusTransportInfrastructure.cs
@@ -55,9 +55,6 @@
 
         public override TransportReceiveInfrastructure ConfigureReceiveInfrastructure()
         {
-            var t = topology as ForwardingTopology;
-            t?.FindOutHowManyTopicExistsForBundle();
-
             return new TransportReceiveInfrastructure(
                 topology.GetMessagePumpFactory(),
                 topology.GetQueueCreatorFactory(),
