+++ src/Transport/NServiceBus.AzureServiceBus.csproj
@@ -131,7 +131,7 @@
     <Compile Include="Topology\MetaModel\NamespaceBundleInfo.cs" />
     <Compile Include="Topology\MetaModel\NamespacePurpose.cs" />
     <Compile Include="Utils\AsyncLazy.cs" />
-    <Compile Include="Topology\PreStartupChecks\NumberOfTopicsInBundleCheck.cs" />
+    <Compile Include="Topology\NumberOfTopicsInBundleCheck.cs" />
     <Compile Include="Utils\MessageReceiverExtensions.cs" />
     <Compile Include="Utils\TypeExtensions.cs" />
     <Compile Include="Topology\TypesScanner\AssemblyTypesScanner.cs" />
@@ -260,6 +260,9 @@
   </ItemGroup>
   <ItemGroup>
     <Analyzer Include="..\packages\Particular.CodeRules.0.1.1\tools\..\analyzers\dotnet\cs\Particular.CodeRules.dll" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Topology\PreStartupChecks\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
