+++ src/Transport/Topology/Topologies/ForwardingTopology.cs
@@ -3,7 +3,6 @@ namespace NServiceBus
     using System;
     using System.Threading.Tasks;
     using AzureServiceBus;
-    using AzureServiceBus.Topology.MetaModel;
     using Logging;
     using Routing;
     using Settings;
@@ -108,18 +107,6 @@ namespace NServiceBus
         public Func<ICreateQueues> GetQueueCreatorFactory()
         {
             return () => container.Resolve<ICreateQueues>();
-        }
-
-        internal void FindOutHowManyTopicExistsForBundle()
-        {
-            var settings = container.Resolve<ReadOnlySettings>();
-            var manageNamespaceManagerLifeCycle = container.Resolve<IManageNamespaceManagerLifeCycle>();
-            var namespaceConfigurations = settings.Get<NamespaceConfigurations>(WellKnownConfigurationKeys.Topology.Addressing.Namespaces);
-            var namespaceBundleConfigurations = settings.Get<NamespaceBundleConfigurations>(WellKnownConfigurationKeys.Topology.Bundling.NamespaceBundleConfigurations);
-            var bundlePrefix = settings.Get<string>(WellKnownConfigurationKeys.Topology.Bundling.BundlePrefix);
-
-            var check = new NumberOfTopicsInBundleCheck(manageNamespaceManagerLifeCycle, namespaceConfigurations, namespaceBundleConfigurations, bundlePrefix);
-            check.Run().GetAwaiter().GetResult();
         }
 
         public Func<IPushMessages> GetMessagePumpFactory()
