+++ src/Transport/Topology/Topologies/ForwardingTopologySectionManager.cs
@@ -17,11 +17,21 @@ namespace NServiceBus.Transport.AzureServiceBus
         readonly ConcurrentDictionary<string, TopologySection> sendDestinations = new ConcurrentDictionary<string, TopologySection>();
         readonly ConcurrentDictionary<Type, TopologySection> publishDestinations = new ConcurrentDictionary<Type, TopologySection>();
         readonly List<EntityInfo> topics = new List<EntityInfo>();
+        Lazy<NamespaceBundleConfigurations> namespaceBundleConfigurations;
 
         public ForwardingTopologySectionManager(SettingsHolder settings, ITransportPartsContainer container)
         {
             this.settings = settings;
             this.container = container;
+
+            namespaceBundleConfigurations = new Lazy<NamespaceBundleConfigurations>(() =>
+            {
+                var manageNamespaceManagerLifeCycle = container.Resolve<IManageNamespaceManagerLifeCycle>();
+                var namespaceConfigurations = settings.Get<NamespaceConfigurations>(WellKnownConfigurationKeys.Topology.Addressing.Namespaces);
+                var bundlePrefix = settings.Get<string>(WellKnownConfigurationKeys.Topology.Bundling.BundlePrefix);
+                var bundleConfigurations = NumberOfTopicsInBundleCheck.Run(manageNamespaceManagerLifeCycle, namespaceConfigurations, bundlePrefix).GetAwaiter().GetResult();
+                return bundleConfigurations;
+            });
         }
 
         public TopologySection DetermineReceiveResources(string inputQueue)
@@ -271,12 +281,11 @@ namespace NServiceBus.Transport.AzureServiceBus
         void BuildTopicBundles(RuntimeNamespaceInfo[] namespaces, AddressingLogic addressingLogic)
         {
             var numberOfEntitiesInBundle = settings.Get<int>(WellKnownConfigurationKeys.Topology.Bundling.NumberOfEntitiesInBundle);
-            var namespaceBundleConfigurations = settings.Get<NamespaceBundleConfigurations>(WellKnownConfigurationKeys.Topology.Bundling.NamespaceBundleConfigurations);
             var bundlePrefix = settings.Get<string>(WellKnownConfigurationKeys.Topology.Bundling.BundlePrefix);
 
             foreach (var @namespace in namespaces)
             {
-                var numberOfTopicsFound = namespaceBundleConfigurations.GetNumberOfTopicInBundle(@namespace.Alias);
+                var numberOfTopicsFound = namespaceBundleConfigurations.Value.GetNumberOfTopicInBundle(@namespace.Alias);
                 var numberOfTopicsToCreate = Math.Max(numberOfEntitiesInBundle, numberOfTopicsFound);
                 for (var i = 1; i <= numberOfTopicsToCreate; i++)
                 {
